[TASK] Update Fluid template files
TYPO3 v9 does not feature the typical four columns in the backend
anymore. Therefore, the Fluid template files have been updated to only
reflect one column.

In addition, the "brand" in the top navigation menu has been updated
to read "TYPO3" and to show the TYPO3 logo. The brand now links to the
website root: "/".